pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "constellation"
version = "2.0.0"
description = "Constellation multi-agent system"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "live: marks tests as requiring live external services (Jira, SCM, Figma, LLM)",
]

[tool.setuptools.packages.find]
include = ["framework*", "agents*", "plugins*"]
